+++ src/services/client-action/imageActions.ts
@@ -1,5 +1,4 @@
-import { formatDateToNumber } from '@/utils/dateUtils';
-import { removeFileExtension } from '@/utils/fileNameUtils';
+import { formatDateToNumber, removeFileExtension } from '@/utils/fileNameUtils';
 import browserClient from '@/utils/supabase/client';
 
 /**
