--- src/utils/getRandomElement.ts
@@ -1,4 +1,0 @@
-export const getRandomElement = (array: any[]) => {
-  const randomIndex = Math.floor(Math.random() * array.length);
-  return array[randomIndex];
-};
