+++ tsconfig.json
@@ -29,7 +29,9 @@
     ".next/types/**/*.ts",
     "src/app/group/[groupId]/post/page.bak",
     "src/app/group/[groupId]/post/page._tsx",
-    "src/app/group/[groupId]/post/ImageUpload.bak"
+    "src/app/group/[groupId]/post/ImageUpload.bak",
+    "src/hooks/queries/byUse/useImageMutations.bak",
+    "src/hooks/queries/byUse/useImageQuery.bak"
   ],
   "exclude": ["node_modules"]
 }
